+++ src/slurmd/slurmd/get_mach_stat.h
@@ -42,9 +42,9 @@
 #include <inttypes.h>
 
 extern int get_cpu_load(uint32_t *cpu_load);
-extern int get_free_mem(uint32_t *free_mem);
+extern int get_free_mem(uint64_t *free_mem);
 extern int get_mach_name(char *node_name);
-extern int get_memory(uint32_t *real_memory);
+extern int get_memory(uint64_t *real_memory);
 extern int get_tmp_disk(uint32_t *tmp_disk, char *tmp_fs);
 extern int get_up_time(uint32_t *up_time);
