+++ src/slurmd/slurmd/slurmd.h
@@ -92,7 +92,7 @@ typedef struct slurmd_config {
 	uint16_t     actual_sockets;    /* actual sockets count            */
 	uint16_t     actual_cores;      /* actual core count               */
 	uint16_t     actual_threads;    /* actual thread per core count    */
-	uint32_t     real_memory_size;  /* amount of real memory	   */
+	uint64_t     real_memory_size;  /* amount of real memory	   */
 	uint32_t     tmp_disk_space;    /* size of temporary disk	   */
 	uint32_t     up_time;		/* seconds since last boot time    */
 	uint16_t     block_map_size;	/* size of block map               */
